task 2 works correctly

+++ more_pointers_arrays_and_strings/2-strncpy.c
@@ -1,5 +1,4 @@
 #include "main.h"
-#include <stddef.h>
 
 /**
 * _strncpy - copies a string
@@ -13,7 +12,7 @@
 char *_strncpy(char *dest, char *src, int n)
 {
 
-	size_t i;
+	int i;
 
 	for (i = 0; i < n && src[i] != '\0'; i++)
 	{
